move network parsing to shared helpers
