+++ smart-contracts/hardhat.config.js
@@ -1,5 +1,6 @@
 // hardhat.config.js
 const { copySync } = require('fs-extra')
+const { networks } = require('@unlock-protocol/hardhat-helpers')
 
 require('@nomiclabs/hardhat-ethers')
 require('@nomiclabs/hardhat-truffle5')
@@ -34,8 +35,6 @@ require('@unlock-protocol/hardhat-plugin')
 // import helpers
 const { etherscan } = require('@unlock-protocol/hardhat-helpers')
 
-const { getHardhatNetwork } = require('./helpers/network')
-
 const settings = {
   optimizer: {
     enabled: true,
@@ -48,7 +47,7 @@ const settings = {
   },
 }
 
-const networks = getHardhatNetwork()
+// used for tests
 networks.hardhat = {
   initialBaseFeePerGas: 100000000,
 }
@@ -84,9 +83,7 @@ require('./tasks/balance')
 require('./tasks/deploy')
 require('./tasks/upgrade')
 require('./tasks/set')
-require('./tasks/safe')
 require('./tasks/release')
-require('./tasks/gov')
 require('./tasks/utils')
 require('./tasks/lock')
 require('./tasks/verify')
